# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kira Code API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/relay-auth/spake2/start": {
      "post": {
        "tags": [
          "relay-auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSpake2EnrollmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse69"
                }
              }
            }
          }
        }
      }
    },
    "/relay-auth/spake2/finish": {
      "post": {
        "tags": [
          "relay-auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishSpake2EnrollmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse70"
                }
              }
            }
          }
        }
      }
    },
    "/relay-auth/signing-session/refresh": {
      "post": {
        "tags": [
          "relay-auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRelaySigningSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse71"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StartSpake2EnrollmentRequest": {
        "type": "object",
        "properties": {
          "enrollment_code": {
            "type": "string"
          },
          "client_message_b64": {
            "type": "string"
          }
        },
        "required": [
          "enrollment_code",
          "client_message_b64"
        ]
      },
      "ApiResponse69": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StartSpake2EnrollmentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StartSpake2EnrollmentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "StartSpake2EnrollmentResponse": {
        "type": "object",
        "properties": {
          "enrollment_id": {
            "type": "string",
            "format": "uuid"
          },
          "server_message_b64": {
            "type": "string"
          }
        },
        "required": [
          "enrollment_id",
          "server_message_b64"
        ]
      },
      "FinishSpake2EnrollmentRequest": {
        "type": "object",
        "properties": {
          "enrollment_id": {
            "type": "string",
            "format": "uuid"
          },
          "client_id": {
            "type": "string",
            "format": "uuid"
          },
          "client_name": {
            "type": "string"
          },
          "client_browser": {
            "type": "string"
          },
          "client_os": {
            "type": "string"
          },
          "client_device": {
            "type": "string"
          },
          "public_key_b64": {
            "type": "string"
          },
          "client_proof_b64": {
            "type": "string"
          }
        },
        "required": [
          "enrollment_id",
          "client_id",
          "client_name",
          "client_browser",
          "client_os",
          "client_device",
          "public_key_b64",
          "client_proof_b64"
        ]
      },
      "ApiResponse70": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinishSpake2EnrollmentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinishSpake2EnrollmentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "FinishSpake2EnrollmentResponse": {
        "type": "object",
        "properties": {
          "signing_session_id": {
            "type": "string",
            "format": "uuid"
          },
          "server_public_key_b64": {
            "type": "string"
          },
          "server_proof_b64": {
            "type": "string"
          }
        },
        "required": [
          "signing_session_id",
          "server_public_key_b64",
          "server_proof_b64"
        ]
      },
      "RefreshRelaySigningSessionRequest": {
        "type": "object",
        "properties": {
          "client_id": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          },
          "nonce": {
            "type": "string"
          },
          "signature_b64": {
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "timestamp",
          "nonce",
          "signature_b64"
        ]
      },
      "ApiResponse71": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RefreshRelaySigningSessionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RefreshRelaySigningSessionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "RefreshRelaySigningSessionResponse": {
        "type": "object",
        "properties": {
          "signing_session_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "signing_session_id"
        ]
      },
...
    }
  }
}